# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "所持ゲーム管理システムのバックエンドAPI仕様書",
        "title": "Yutagame Backend API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/admin/games": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "条件に従いゲーム情報の一覧を取得します。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Games"
                ],
                "summary": "ゲーム一覧取得",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ページ番号 (指定するとページングモード)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "表示件数 (10, 30, 50)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "自由入力のテキスト検索（ゲーム名）",
                        "name": "q",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "collectionFormat": "csv",
                        "description": "メーカーIDの複数指定",
                        "name": "manufacturerIDs",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "collectionFormat": "csv",
                        "description": "機種IDの複数指定",
                        "name": "machineIDs",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "collectionFormat": "csv",
                        "description": "ジャンルIDの複数指定",
                        "name": "genreIDs",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "collectionFormat": "csv",
                        "description": "キーワードIDの複数指定",
                        "name": "keywordIDs",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "page指定時",
                        "schema": {
                            "$ref": "#/definitions/handler.PaginatedResponse-model_Game"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "新しいゲームの情報をデータベースに登録します。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Games"
                ],
                "summary": "ゲーム新規登録",
                "parameters": [
                    {
                        "description": "ゲーム登録情報",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/admin.GameSaveRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.Game"
                        }
                    }
                }
            }
        },
...
        "/admin/games/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "指定されたIDまたはコードのゲーム詳細情報を取得します。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Games"
                ],
                "summary": "ゲーム詳細取得",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ゲームIDまたはゲームコード",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Game"
                        }
                    },
                    "404": {
                        "description": "未検出エラー",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "指定されたIDのゲーム情報更新します。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Games"
                ],
                "summary": "ゲーム情報更新",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ゲームID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "ゲーム更新情報",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/admin.GameSaveRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Game"
                        }
                    },
                    "400": {
                        "description": "エラー",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "指定されたIDのゲーム情報を削除します。",
                "tags": [
                    "Games"
                ],
                "summary": "ゲーム情報削除",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ゲームID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "エラー",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/app/games": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Public Games"
                ],
                "summary": "ゲーム検索",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ゲーム名・カナ",
                        "name": "q",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "機種コード",
                        "name": "machineCode",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "ジャンルコード",
                        "name": "genreCode",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "メーカーコード",
                        "name": "manufacturerCode",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "キーワードコード",
                        "name": "keywordCode",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "並び順",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "ページ番号",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "表示件数",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "匿名訪問者ID",
                        "name": "visitorId",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.PaginatedGameResponse"
                        }
                    }
                }
            }
        },
        "/app/games/{code}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Public Games"
                ],
                "summary": "ゲーム詳細取得",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ゲームコード",
                        "name": "code",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "匿名訪問者ID",
                        "name": "visitorId",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Game"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/app/page-views": {
            "post": {
                "description": "同じ訪問者による同じページの閲覧は、日本時間で1日1回だけPVとして記録します。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Analytics"
                ],
                "summary": "ページ閲覧を日次記録",
                "parameters": [
                    {
                        "description": "ページ閲覧情報",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/app.PageViewRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/app/rankings/page": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Public Rankings"
                ],
                "summary": "集計ランキング取得",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ランキング種別",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "集計期間",
                        "name": "period",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "ページ番号",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "表示件数",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.PublicRankingResponse"
                        }
                    }
                }
            }
        },
...
        "/app/top": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Public Games"
                ],
                "summary": "TOP画面コンテンツ取得",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 5,
                        "description": "発売日枠の件数",
                        "name": "releaseLimit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 8,
                        "description": "新着枠の件数",
                        "name": "recentLimit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 8,
                        "description": "ランダム・お気に入り枠の件数",
                        "name": "randomLimit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.TopContents"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "admin.GameSaveRequest": {
            "type": "object",
            "properties": {
                "catchCopy": {
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "genreId": {
                    "type": "integer"
                },
                "isClear": {
                    "type": "boolean"
                },
                "isFavourite": {
                    "type": "boolean"
                },
                "isPlay": {
                    "type": "boolean"
                },
                "kana": {
                    "type": "string"
                },
                "keywordIds": {
                    "description": "💡 紐付けるキーワードのID配列",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "listPrice": {
                    "type": "integer"
                },
                "machineId": {
                    "type": "integer"
                },
                "manufacturerId": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "officialSiteUrl": {
                    "type": "string"
                },
                "overview": {
                    "type": "string"
                },
                "releaseDate": {
                    "description": "フロントからは文字列(\"YYYY-MM-DD\")で受け取る",
                    "type": "string"
                },
                "subCatch": {
                    "type": "string"
                },
                "subGenre": {
                    "type": "string"
                },
                "youtubeUrl": {
                    "type": "string"
                }
            }
        },
...
        "app.PageViewRequest": {
            "type": "object",
            "properties": {
                "pagePath": {
                    "type": "string"
                },
                "visitorId": {
                    "type": "string"
                }
            }
        },
        "app.PaginatedGameResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "totalCount": {
                    "type": "integer"
                },
                "totalPages": {
                    "type": "integer"
                }
            }
        },
        "app.PublicRankingResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Game"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "totalCount": {
                    "type": "integer"
                },
                "totalPages": {
                    "type": "integer"
                }
            }
        },
...
        "app.TopContents": {
            "type": "object",
            "properties": {
                "favoriteRanking": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/app.FavoriteRankingItem"
                    }
                },
                "randomPicks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Game"
                    }
                },
                "rankingTop20": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Game"
                    }
                },
                "recentlyReleased": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Game"
                    }
                },
                "recentlyUpdated": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Game"
                    }
                },
                "releaseToday": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Game"
                    }
                }
            }
        },
        "handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "handler.PaginatedResponse-model_Game": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "admins や items などのキー名に依存しないよう、汎用的に \"data\" とします",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Game"
                    }
                },
                "totalCount": {
                    "type": "integer"
                },
                "totalPages": {
                    "type": "integer"
                }
            }
        },
...
        "model.Game": {
            "type": "object",
            "properties": {
                "affiliates": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.GameAffiliate"
                    }
                },
                "catchCopy": {
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "genre": {
                    "$ref": "#/definitions/model.Genre"
                },
                "genreId": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "imageKey": {
                    "type": "string"
                },
                "isClear": {
                    "type": "boolean"
                },
                "isFavourite": {
                    "type": "boolean"
                },
                "isPlay": {
                    "type": "boolean"
                },
                "kana": {
                    "type": "string"
                },
                "keywords": {
                    "description": "💡 凄まじいポイント：GORMに「game_keywordsという中間テーブルを使って、紐づくKeywordsを全自動でガッチャンコしてね」と1行で命令します",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Keyword"
                    }
                },
                "listPrice": {
                    "type": "integer"
                },
                "machine": {
                    "$ref": "#/definitions/model.Machine"
                },
                "machineId": {
                    "type": "integer"
                },
                "manufacturer": {
                    "description": "💡 依存関係の解消：GORMの1対多（Belongs To）リレーション群",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Manufacturer"
                        }
                    ]
                },
                "manufacturerId": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "officialSiteUrl": {
                    "type": "string"
                },
                "overview": {
                    "type": "string"
                },
                "previousRank": {
                    "type": "integer"
                },
                "rank": {
                    "type": "integer"
                },
                "rankingCount": {
                    "type": "integer"
                },
                "releaseDate": {
                    "type": "string"
                },
                "subCatch": {
                    "type": "string"
                },
                "subGenre": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "youtubeUrl": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "\"Bearer {token}\" の形式でJWTトークンを入力してください。",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}